refactor: 산책로 mapper 삭제
