+++ app-external-api/src/main/java/com/dongsan/api/domains/walkway/dto/request/UpdateWalkwayRequest.java
@@ -1,6 +1,7 @@
 package com.dongsan.api.domains.walkway.dto.request;
 
 import com.dongsan.core.domains.walkway.ExposeLevel;
+import com.dongsan.core.domains.walkway.UpdateWalkway;
 import jakarta.validation.constraints.NotNull;
 import java.util.List;
 
@@ -13,4 +14,13 @@ public record UpdateWalkwayRequest(
         List<String> hashtags,
         ExposeLevel exposeLevel
 ) {
+        public UpdateWalkway toUpdateWalkway(Long walkwayId) {
+                return new UpdateWalkway(
+                        walkwayId,
+                        this.name(),
+                        this.memo(),
+                        this.exposeLevel(),
+                        this.hashtags()
+                );
+        }
 }
