+++ app-external-api/src/main/java/com/dongsan/api/domains/walkway/dto/response/GetWalkwayResponse.java
@@ -1,7 +1,7 @@
 package com.dongsan.api.domains.walkway.dto.response;
 
 import com.dongsan.api.domains.walkway.dto.WalkwayCoordinate;
-import com.dongsan.api.domains.walkway.mapper.LineStringMapper;
+import com.dongsan.api.domains.walkway.LineStringMapper;
 import com.dongsan.core.domains.walkway.Walkway;
 import com.dongsan.core.domains.walkway.ExposeLevel;
 import java.time.format.DateTimeFormatter;
